取消代码文件读取锁

+++ TextLocator/Index/IndexCore.cs
@@ -202,7 +202,7 @@ namespace TextLocator.Index
                     indexWriter.Optimize();
                 }
 
-                string msg = "索引：[" + finishCount * 1.0F + "/" + taskInfo.TotalCount + "] => 引擎：" + fileType + "，文件：" + filePath + "，耗时：" + workMark.TotalSeconds + "秒";
+                string msg = "索引：[" + finishCount * 1.0F + "/" + taskInfo.TotalCount + "] => 引擎：" + (int)fileType + "，文件：" + filePath + "，耗时：" + workMark.TotalSeconds + "秒";
 
                 // 执行状态回调
                 taskInfo.Callback(msg, CalcCompletionRatio(finishCount, taskInfo.TotalCount));
